SignLanguageConvEnv/SignLanConverter/requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
decorator==4.4.2
Django==4.1.7
djangorestframework==3.14.0
docopt==0.6.2
ffmpeg-python==0.2.0
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
joblib==1.2.0
moviepy==1.0.3
nltk==3.8.1
numpy==1.24.2
Pillow==9.4.0
pipreqs==0.4.11
proglog==0.1.10
psutil==5.9.4
pycodestyle==2.10.0
pycparser==2.21
pydub==0.25.1
pytz==2023.3
regex==2022.10.31
requests==2.28.2
SpeechRecognition==3.10.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
tzdata==2022.7
urllib3==1.26.15
yarg==0.1.9